Added interfaces to event handlers
